Move StorageComponentContainerContributor to extensions package
